refactor: encapsule message builders for Cyclopedia, Bosstiary and others
Introduces new message builder classes for Cyclopedia Character, House, Outfit, Imbuement, Prey, and Bosstiary features in the network protocol. Updates CMakeLists.txt to include these new source files. Refactors Player class to expose marriage spouse ID and item type count methods for protocol use.
